feat(6737): extract topology coherence contract tests
