+++ crates/kamn-node/src/main_tests/daemon_tests/live_postgres_topology_contract_tests/topology_coherence_contract_tests/host_mode_host_pair_contract_tests.rs
@@ -1,0 +1,6 @@
+// include!("live_postgres_topology_contract_tests/topology_coherence_contract_tests/host_mode_host_pair_contract_tests/coherence_contract_tests.rs");
+#[path = "host_mode_host_pair_contract_tests/coherence_contract_tests.rs"]
+mod coherence_contract_tests;
+// include!("live_postgres_topology_contract_tests/topology_coherence_contract_tests/host_mode_host_pair_contract_tests/cardinality_contract_tests.rs");
+#[path = "host_mode_host_pair_contract_tests/cardinality_contract_tests.rs"]
+mod cardinality_contract_tests;
