+++ crates/kamn-node/src/main_tests/daemon_tests/live_postgres_topology_contract_tests/topology_coherence_contract_tests/lane_set_bundle_contract_tests.rs
@@ -1,0 +1,6 @@
+// include!("live_postgres_topology_contract_tests/topology_coherence_contract_tests/lane_set_bundle_contract_tests/lane_count_contract_tests.rs");
+#[path = "lane_set_bundle_contract_tests/lane_count_contract_tests.rs"]
+mod lane_count_contract_tests;
+// include!("live_postgres_topology_contract_tests/topology_coherence_contract_tests/lane_set_bundle_contract_tests/lane_id_bundle_contract_tests.rs");
+#[path = "lane_set_bundle_contract_tests/lane_id_bundle_contract_tests.rs"]
+mod lane_id_bundle_contract_tests;
